refactor message format functions
