+++ lib/checktarget/messageMaker.ts
@@ -19,13 +19,9 @@ import {
     buttonForCommand,
     HandlerContext,
     HandlerResult,
-    logger,
     ParameterType,
 } from "@atomist/automation-client";
 import {
-    consistentHash,
-    Diff,
-    FP,
     Vote,
     VoteResults,
 } from "@atomist/clj-editors";
@@ -42,15 +38,14 @@ import {
     SlackMessage,
 } from "@atomist/slack-messages";
 import * as _ from "lodash";
-import { Feature } from "../..";
 import { toName } from "../adhoc/preferences";
 import {
     ApplyAllFingerprintsName,
     ApplyTargetFingerprintName,
 } from "../handlers/commands/applyFingerprint";
 import { UpdateTargetFingerprintName } from "../handlers/commands/updateTarget";
-import { DiffSummary } from "../machine/Feature";
-import { applyToFeature, displayName, displayValue } from "../machine/Features";
+import { applyFingerprintTitle, GitCoordinate, prBody } from "../support/messages";
+import { orDefault } from "../support/util";
 
 export interface MessageMakerParams {
     ctx: HandlerContext;
@@ -60,72 +55,7 @@ export interface MessageMakerParams {
     coord: GitCoordinate;
 }
 
-export interface GitCoordinate {
-    owner: string;
-    repo: string;
-    sha: string;
-    providerId: string;
-    branch?: string;
-}
-
 export type MessageMaker = (params: MessageMakerParams) => Promise<HandlerResult>;
-
-type MessageIdMaker = (shas: string[], coordinate: GitCoordinate, channel: string) => string;
-
-export const updateableMessage: MessageIdMaker = (shas, coordinate: GitCoordinate, channel: string) => {
-    return consistentHash([...shas, channel, coordinate.owner, coordinate.repo]);
-    // return _.times(20, () => _.random(35).toString(36)).join("");
-};
-
-/**
- * get a diff summary if any registrations support one for this Fingerprint type
- */
-export function getDiffSummary(diff: Diff, target: FP, feature: Feature): undefined | DiffSummary {
-
-    try {
-        if (feature.summary) {
-            return feature.summary(diff, target);
-        } else {
-            return {
-                title: "Target diff",
-                description: `from ${diff.from.data} to ${diff.to.data}`,
-            };
-        }
-    } catch (e) {
-        logger.warn(`failed to create summary: ${e}`);
-    }
-
-    return undefined;
-}
-
-function orDefault<T>(cb: () => T, x: T): T {
-    try {
-        return cb();
-    } catch (y) {
-        return x;
-    }
-}
-
-function applyFingerprintTitle(vote: Vote): string {
-    try {
-        return `Apply fingerprint ${applyToFeature(vote.fpTarget, displayName)} (${applyToFeature(vote.fpTarget, displayValue)})`;
-    } catch (ex) {
-        return `Apply fingerprint ${vote.fpTarget.name}`;
-    }
-}
-
-function prBody(vote: Vote): string {
-    const title: string =
-        orDefault(
-            () => vote.summary.title,
-            applyFingerprintTitle(vote));
-    const description: string =
-        orDefault(
-            () => vote.summary.description,
-            `no summary`);
-
-    return `#### ${title}\n${description}`;
-}
 
 /**
  * Message for one case where a Fingerprint target is different from what's in the latest Push.
@@ -148,8 +78,9 @@ function oneFingerprint(params: MessageMakerParams, vote: Vote): Attachment {
                 {
                     msgId: params.msgId,
                     targetfingerprint: toName(vote.fpTarget.type, vote.fpTarget.name),
-                    title: applyFingerprintTitle(vote),
+                    title: applyFingerprintTitle(vote.fpTarget),
                     branch: vote.diff.branch,
+                    body: prBody(vote),
                     targets: {
                         owner: vote.diff.owner,
                         repo: vote.diff.repo,
